refactor : 파일 분리
- bar -> bar, sign up bar, login bar
